let's change to url_for in the redirect
The bare `redirect(username)` turns out to be confusing to students and error prone, so I suggest changing that to an explicit `url_for` which passes the username as a parameter to the relevant view function.

I'm sorry and ashamed to say that this bare redirect came from my original code :(

I know that now it'll be annoying to change, since it appears (in some slight variation) in all parts of the FlaskChatApplication repos, but I think it's absolutely worth including in the next update to this lesson. Also, to make it easier to maintain, I'd much prefer to combine all of these into a single large repo.